+++ admin/pages/pets.php
@@ -17,9 +17,13 @@
         exit;
     }
 
-
+//search the pet based on bread and gives the details accordingly
+if(isset($_GET['search'])) {
+    $breed = strtoupper($_GET["search"]);
+    $sql = "SELECT * FROM PETS WHERE BREED='".$breed."' ORDER BY NAME";
+} else {
     $sql = "SELECT * FROM PETS ORDER BY NAME";
-
+}
     $result = mysqli_query($conn,$sql);
 ?>
 
@@ -48,16 +52,32 @@
         </div>
     </nav>
 
+    <div class="row">
+        <div class="cols-sm-3"></div>
+        <div class="cols-sm-3"></div>
+        <div class="cols-sm-3">
+            <form action="pets.php" method="get">
+                <input type="search" name="search" placeholder="Enter Breed">
+                <button type="submit" class="btn btn-danger">Search Pets</button>
+            </form>
+        </div>
+        <div class="cols-sm-3"></div>
+    </div>
+
     <?php
-    echo "<table class='table table-responsive table-bordered table-striped'><tr class='table-header'><th>NAME</th><th>BREED</th><th>AGE</th><th>PRICE</th><th>EDIT</th><th>DELETE</th></tr>";
-    // output data of each row
-    while($row = $result->fetch_assoc()) {
-        echo "<tr><td>".$row["NAME"]."</td><td>".$row["BREED"]."</td><td>".$row["AGE"]."</td><td>$ ".$row["PRICE"]."</td>
-         <td><a href='edit_pet.php?id=".$row['PET_ID']."'> <i class='fa fa-edit'>&nbsp</i></a></td>
-          <td><a href='../processes/delete_pet.php?id=".$row['PET_ID']."'> <i class='fa fa-remove'>&nbsp</i></a></td>
+    if($result->num_rows == 0) {
+        echo "<div class='alert alert-danger'>No Record Found</div>";
+    } else {
+        echo "<table class='table table-responsive table-bordered table-striped'><tr class='table-header'><th>NAME</th><th>BREED</th><th>AGE</th><th>PRICE</th><th>EDIT</th><th>DELETE</th></tr>";
+        // output data of each row
+        while ($row = $result->fetch_assoc()) {
+            echo "<tr><td>" . $row["NAME"] . "</td><td>" . $row["BREED"] . "</td><td>" . $row["AGE"] . "</td><td>$ " . $row["PRICE"] . "</td>
+         <td><a href='edit_pet.php?id=" . $row['PET_ID'] . "'> <i class='fa fa-edit'>&nbsp</i></a></td>
+          <td><a href='../processes/delete_pet.php?id=" . $row['PET_ID'] . "'> <i class='fa fa-remove'>&nbsp</i></a></td>
           </tr>";
+        }
+        echo "</table>";
     }
-    echo "</table>";
     ?>
 </div>
